feat: rename project to skill-guard (closes #24)
- PyPI package: skill-guard
- CLI command: skill-guard
- Python package dir: skill_gate/ → skill_guard/
- All imports, docs, workflows, README updated
- Version bumped to 0.4.0
- 64 tests passing, 81.62% coverage
