+++ skill_guard/commands/validate.py
@@ -1,4 +1,4 @@
-"""CLI command: skill-gate secure"""
+"""CLI command: skill-gate validate"""
 
 from __future__ import annotations
 
@@ -6,31 +6,33 @@ from pathlib import Path
 
 import typer
 
-from skill_gate.config import load_config
-from skill_gate.engine.security import run_security_scan
-from skill_gate.models import ConfigError, SkillParseError
-from skill_gate.output.json_out import format_as_json
-from skill_gate.output.markdown import format_as_markdown
-from skill_gate.output.text import format_security_result
-from skill_gate.parser import parse_skill
+from skill_guard.config import load_config
+from skill_guard.engine.quality import run_validation
+from skill_guard.models import ConfigError, SkillParseError
+from skill_guard.output.json_out import format_as_json
+from skill_guard.output.markdown import format_as_markdown
+from skill_guard.output.text import format_validation_result
+from skill_guard.parser import parse_skill
 
 SKILL_PATH_ARG = typer.Argument(..., help="Path to skill directory")
 CONFIG_PATH_OPT = typer.Option(None, "--config", help="Path to skill-gate.yaml")
 FORMAT_OPT = typer.Option("text", "--format", help="Output format: text|json|md")
 QUIET_OPT = typer.Option(False, "--quiet", help="Suppress non-essential output")
+VERBOSE_OPT = typer.Option(False, "--verbose", help="Show all check details")
 
 
-def secure_cmd(
+def validate_cmd(
     skill_path: Path = SKILL_PATH_ARG,
     config_path: Path | None = CONFIG_PATH_OPT,
     format: str = FORMAT_OPT,
     quiet: bool = QUIET_OPT,
+    verbose: bool = VERBOSE_OPT,
 ):
-    """Scan a skill for dangerous patterns."""
+    """Validate a skill against format and quality rules."""
     try:
         config = load_config(config_path)
         skill = parse_skill(skill_path)
-        result = run_security_scan(skill, config.secure)
+        result = run_validation(skill, config.validate)
     except ConfigError as e:
         typer.echo(f"Config error: {e}")
         raise typer.Exit(code=3) from e
@@ -39,11 +41,13 @@ def secure_cmd(
         raise typer.Exit(code=4) from e
 
     if format == "json":
-        typer.echo(format_as_json(result, command="secure"))
+        typer.echo(format_as_json(result, command="validate"))
     elif format in ("md", "markdown"):
-        typer.echo(format_as_markdown(result, command="secure"))
+        typer.echo(format_as_markdown(result, command="validate"))
     else:
-        format_security_result(result, quiet=quiet)
+        format_validation_result(result, quiet=quiet, verbose=verbose)
 
-    if not result.passed:
+    if result.blockers > 0:
         raise typer.Exit(code=1)
+    if result.warnings > 0:
+        raise typer.Exit(code=2)
