+++ tests/unit/test_output.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 
-from skill_gate.models import CheckResult, ConflictResult, ValidationResult
-from skill_gate.output.json_out import format_as_json
-from skill_gate.output.markdown import format_as_markdown
+from skill_guard.models import CheckResult, ConflictResult, ValidationResult
+from skill_guard.output.json_out import format_as_json
+from skill_guard.output.markdown import format_as_markdown
 
 
 def test_json_output_roundtrip():
